voxform-fastapi/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-jose[cryptography]>=3.3
bcrypt>=4.0
databases[aiomysql]>=0.9
aiomysql>=0.2
python-multipart>=0.0.9
httpx>=0.27
python-dotenv>=1.0
